Re-ordered client
Put all client files in Client folder
